Rename containers in documenation
