+++ packages/plugins/documentation/admin/src/pages/App/index.js
@@ -6,23 +6,14 @@
  */
 
 import React from 'react';
-import { Switch, Route } from 'react-router-dom';
-import { NotFound, CheckPagePermissions } from '@strapi/helper-plugin';
-// Utils
+import { CheckPagePermissions } from '@strapi/helper-plugin';
 import pluginPermissions from '../../permissions';
-import pluginId from '../../pluginId';
-// Containers
 import HomePage from '../HomePage';
 
 function App() {
   return (
     <CheckPagePermissions permissions={pluginPermissions.main}>
-      <div className={pluginId}>
-        <Switch>
-          <Route path={`/plugins/${pluginId}`} component={HomePage} exact />
-          <Route component={NotFound} />
-        </Switch>
-      </div>
+      <HomePage />
     </CheckPagePermissions>
   );
 }
