+++ packages/plugins/documentation/admin/src/pages/HomePage/utils/schema.js
